move sidebar button to header
