+++ components/layout/mainLayout.js
@@ -1,6 +1,6 @@
 import AppHeader from 'components/header.jsx'
 import AppFooter from 'components/footer.jsx'
-import Sidebar from 'components/sidebar.jsx';
+// import Sidebar from 'components/sidebar.jsx';
 import { AuthProvider } from 'src/authContext';
 
 export default function Layout({ children }) {
@@ -9,7 +9,7 @@ export default function Layout({ children }) {
         <AuthProvider>
             <AppHeader />
                 <div className="app-container">                
-                    <Sidebar />                    
+                    {/* <Sidebar />                     */}
                     <main>
                         { children }
                     </main>
